rename useServers to useCentralServers
